% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bms.R
\name{bms}
\alias{bms}
\title{Bayesian Model Confidence Set}
\usage{
bms(data, alpha, eps = 1e-06)
}
\arguments{
\item{data}{a list including
\describe{
\item{x}{covariates matrix, of dimension nobs and nvars;each row is an observation vector.}
\item{y}{response variable.}
}}

\item{alpha}{a vector of significance levels. The confidence levels are 1-\code{alpha}.
Default value is 0.05.}

\item{eps}{toterance level in choosing models with total posteriors
at least \code{1-alpha}. Default value is \code{1e-6}.}
}
\value{
Returns a list containing:
\item{models}{A list with one entry for each model. Each entry is an integer
vector that specifies the columns of matrix \code{x} to be used as a regressor in that model.
Models is ordered with decreasing posteriors.}
\item{con_sets}{a list with with one entry for a \code{1-alpha} model confidence set.
Each entry is an integer vector that specifies the models selected in this set. The model
indexes used in \code{con_sets} are their orders in \code{models}.}
\item{length_con}{lengths of confidence sets.}
\item{probs_inorder}{Model posteriors in decreasing order.}
\item{beta_ls}{a list with one entry for each model. Each entry is a vector
of estimated coefficients for that model.}
}
\description{
This function allows you to obtain a bayesian model confidence set with
prosterior as the confidence level of each model.
}
\examples{
n= 100
B=100
p=6
x = matrix(rnorm(n*p, mean=0, sd=1), n, p)
true_b = c(1:3, rep(0,p-3))
y = x\%*\% true_b+rnorm(n)
alpha=c(0.1,0.05,0.01)
data=list(x=x,y=y)
result=bms(data,alpha)
}
\keyword{Bayesian}
