% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_gen.R
\name{Models_gen}
\alias{Models_gen}
\title{Generate all subset models}
\usage{
Models_gen(predictors)
}
\arguments{
\item{predictors}{a vector including the indexes of all predictors,
such as \code{1:p}.}
}
\value{
Returns a list with one entry for each model. Each entry is an integer
vector that specifies the columns of matrix \code{x} to be used as a regressor in that model.
}
\description{
This function generates a list including all subset models given a
vector of candidate predictors.
}
\examples{
Models_gen(1:5)
}
\seealso{
\code{\link[utils]{combn}}, \code{\link[rlist]{list.flatten}}
}
